Fix a DNF check
Return statements in an eval block, in this case the "return $? == 0;"
statement, are evaluated in a context of the eval keyword. In this case
it's a void context and thus the return value gets discarded. Then the
final return after eval always returns false.

As a result the _detect_dnf() never worked. This patch fixes it.